+++ __config__/jest.d.ts
@@ -1,7 +1,7 @@
 /**
  * This file is part of Serlo.org API
  *
- * Copyright (c) 2020-2022 Serlo Education e.V.
+ * Copyright (c) 2020-2023 Serlo Education e.V.
  *
  * Licensed under the Apache License, Version 2.0 (the "License")
  * you may not use this file except in compliance with the License
@@ -15,14 +15,15 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  *
- * @copyright Copyright (c) 2020-2022 Serlo Education e.V.
+ * @copyright Copyright (c) 2020-2023 Serlo Education e.V.
  * @license   http://www.apache.org/licenses/LICENSE-2.0 Apache License 2.0
  * @link      https://github.com/serlo-org/api.serlo.org for the canonical source repository
  */
 import { Event } from '@sentry/node'
 
-import { MockTimer } from './setup'
-import { Cache } from '~/internals/cache'
+import type { MockKratos } from '../__tests__/__utils__/services'
+import type { MockTimer } from './setup'
+import type { Cache } from '~/internals/cache'
 
 export {}
 
@@ -32,4 +33,5 @@ declare global {
   var server: ReturnType<typeof import('msw/node').setupServer>
   var timer: MockTimer
   var sentryEvents: Event[]
+  var kratos: MockKratos
 }