+++ __config__/setup.ts
@@ -1,7 +1,7 @@
 /**
  * This file is part of Serlo.org API
  *
- * Copyright (c) 2020-2022 Serlo Education e.V.
+ * Copyright (c) 2020-2023 Serlo Education e.V.
  *
  * Licensed under the Apache License, Version 2.0 (the "License")
  * you may not use this file except in compliance with the License
@@ -15,7 +15,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  *
- * @copyright Copyright (c) 2020-2022 Serlo Education e.V.
+ * @copyright Copyright (c) 2020-2023 Serlo Education e.V.
  * @license   http://www.apache.org/licenses/LICENSE-2.0 Apache License 2.0
  * @link      https://github.com/serlo-org/api.serlo.org for the canonical source repository
  */
@@ -26,6 +26,7 @@ import { setupServer } from 'msw/node'
 import {
   defaultSpreadsheetApi,
   givenSpreadheetApi,
+  MockKratos,
 } from '../__tests__/__utils__'
 import { createCache } from '~/internals/cache'
 import { initializeSentry, Sentry } from '~/internals/sentry'
@@ -64,10 +65,12 @@ export function setup() {
   const timer = new MockTimer()
   const cache = createCache({ timer })
   const server = setupServer()
+  const kratos = new MockKratos()
 
   global.cache = cache
   global.server = server
   global.timer = timer
+  global.kratos = kratos
 }
 
 export async function createBeforeAll(options: SharedOptions) {
@@ -98,6 +101,7 @@ export async function createBeforeEach() {
   await global.cache.flush()
   global.timer.flush()
   global.sentryEvents = []
+  global.kratos.identities = []
 
   process.env.ENVIRONMENT = 'local'
 }