+++ __tests__/schema/user/delete-regular-users.ts
@@ -1,7 +1,7 @@
 /**
  * This file is part of Serlo.org API
  *
- * Copyright (c) 2020-2022 Serlo Education e.V.
+ * Copyright (c) 2020-2023 Serlo Education e.V.
  *
  * Licensed under the Apache License, Version 2.0 (the "License")
  * you may not use this file except in compliance with the License
@@ -15,7 +15,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  *
- * @copyright Copyright (c) 2020-2022 Serlo Education e.V.
+ * @copyright Copyright (c) 2020-2023 Serlo Education e.V.
  * @license   http://www.apache.org/licenses/LICENSE-2.0 Apache License 2.0
  * @link      https://github.com/serlo-org/api.serlo.org for the canonical source repository
  */
@@ -23,16 +23,7 @@ import { gql } from 'apollo-server'
 import * as R from 'ramda'
 
 import { user as baseUser } from '../../../__fixtures__'
-import {
-  Client,
-  given,
-  Database,
-  returnsUuidsFromDatabase,
-  nextUuid,
-  Query,
-} from '../../__utils__'
-
-let database: Database
+import { Client, given, nextUuid, Query } from '../../__utils__'
 
 let client: Client
 let mutation: Query
@@ -43,9 +34,6 @@ const noUserId = nextUuid(nextUuid(user.id))
 
 beforeEach(() => {
   client = new Client({ userId: user.id })
-
-  database = new Database()
-  database.hasUuids(users)
 
   mutation = client
     .prepareQuery({
@@ -66,15 +54,17 @@ beforeEach(() => {
   given('UserDeleteRegularUsersMutation').isDefinedBy((req, res, ctx) => {
     const { userId } = req.body.payload
 
-    database.deleteUuid(userId)
+    given('UuidQuery').withPayload({ id: userId }).returnsNotFound()
 
     return res(ctx.json({ success: true }))
   })
 
-  given('UuidQuery').isDefinedBy(returnsUuidsFromDatabase(database))
+  given('UuidQuery').for(users)
 })
 
 test('runs successfully when mutation could be successfully executed', async () => {
+  expect(global.kratos.identities).toHaveLength(users.length)
+
   await mutation.shouldReturnData({
     user: {
       deleteRegularUsers: [
@@ -83,6 +73,7 @@ test('runs successfully when mutation could be successfully executed', async () 
       ],
     },
   })
+  expect(global.kratos.identities).toHaveLength(users.length - 2)
 })
 
 test('runs partially when one of the mutations failed', async () => {
@@ -92,10 +83,12 @@ test('runs partially when one of the mutations failed', async () => {
     if (userId === user.id)
       return res(ctx.json({ success: false, reason: 'failure!' }))
 
-    database.deleteUuid(userId)
+    given('UuidQuery').withPayload({ id: userId }).returnsNotFound()
 
     return res(ctx.json({ success: true }))
   })
+
+  expect(global.kratos.identities).toHaveLength(users.length)
 
   await mutation.shouldReturnData({
     user: {
@@ -105,6 +98,7 @@ test('runs partially when one of the mutations failed', async () => {
       ],
     },
   })
+  expect(global.kratos.identities).toHaveLength(users.length - 1)
 })
 
 test('fails when username does not match user', async () => {
@@ -149,6 +143,16 @@ test('fails when user does not have role "sysadmin"', async () => {
 
 test('fails when database layer has an internal error', async () => {
   given('UserDeleteRegularUsersMutation').hasInternalServerError()
+
+  await mutation.shouldFailWithError('INTERNAL_SERVER_ERROR')
+
+  expect(global.kratos.identities).toHaveLength(users.length)
+})
+
+test('fails when kratos has an error', async () => {
+  global.kratos.admin.adminDeleteIdentity = () => {
+    throw new Error('Error in kratos')
+  }
 
   await mutation.shouldFailWithError('INTERNAL_SERVER_ERROR')
 })