+++ packages/server/src/internals/cache/index.ts
@@ -1,7 +1,7 @@
 /**
  * This file is part of Serlo.org API
  *
- * Copyright (c) 2020-2022 Serlo Education e.V.
+ * Copyright (c) 2020-2023 Serlo Education e.V.
  *
  * Licensed under the Apache License, Version 2.0 (the "License")
  * you may not use this file except in compliance with the License
@@ -15,7 +15,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  *
- * @copyright Copyright (c) 2020-2022 Serlo Education e.V.
+ * @copyright Copyright (c) 2020-2023 Serlo Education e.V.
  * @license   http://www.apache.org/licenses/LICENSE-2.0 Apache License 2.0
  * @link      https://github.com/serlo-org/api.serlo.org for the canonical source repository
  */
@@ -25,11 +25,11 @@ import Redis from 'ioredis'
 import createMsgpack from 'msgpack5'
 import * as R from 'ramda'
 
+import { createLockManager, LockManager } from './lock-manager'
 import { log } from '../log'
 import { redisUrl } from '../redis-url'
 import { Time, timeToMilliseconds } from '../swr-queue'
 import { Timer } from '../timer'
-import { createLockManager, LockManager } from './lock-manager'
 import { AsyncOrSync } from '~/utils'
 
 const msgpack = (