+++ packages/server/src/internals/swr-queue.ts
@@ -1,7 +1,7 @@
 /**
  * This file is part of Serlo.org API
  *
- * Copyright (c) 2020-2022 Serlo Education e.V.
+ * Copyright (c) 2020-2023 Serlo Education e.V.
  *
  * Licensed under the Apache License, Version 2.0 (the "License")
  * you may not use this file except in compliance with the License
@@ -15,7 +15,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  *
- * @copyright Copyright (c) 2020-2022 Serlo Education e.V.
+ * @copyright Copyright (c) 2020-2023 Serlo Education e.V.
  * @license   http://www.apache.org/licenses/LICENSE-2.0 Apache License 2.0
  * @link      https://github.com/serlo-org/api.serlo.org for the canonical source repository
  */
@@ -25,6 +25,7 @@ import * as t from 'io-ts'
 import reporter from 'io-ts-reporters'
 import * as R from 'ramda'
 
+import { createAuthServices } from './authentication'
 import { Cache, CacheEntry, Priority } from './cache'
 import { isQuery, QuerySpec } from './data-source-helper'
 import { captureErrorEvent } from './error-event'
@@ -79,6 +80,7 @@ export function createSwrQueue({
     environment: {
       cache,
       swrQueue: emptySwrQueue,
+      authServices: createAuthServices(),
     },
   }
   const models = R.values(modelFactories).map((createModel) =>
@@ -173,6 +175,7 @@ export function createSwrQueueWorker({
     environment: {
       cache,
       swrQueue: emptySwrQueue,
+      authServices: createAuthServices(),
     },
   }
   const models = R.values(modelFactories).map((createModel) =>