+++ packages/server/src/schema/thread/resolvers.ts
@@ -1,7 +1,7 @@
 /**
  * This file is part of Serlo.org API
  *
- * Copyright (c) 2020-2022 Serlo Education e.V.
+ * Copyright (c) 2020-2023 Serlo Education e.V.
  *
  * Licensed under the Apache License, Version 2.0 (the "License")
  * you may not use this file except in compliance with the License
@@ -15,7 +15,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  *
- * @copyright Copyright (c) 2020-2022 Serlo Education e.V.
+ * @copyright Copyright (c) 2020-2023 Serlo Education e.V.
  * @license   http://www.apache.org/licenses/LICENSE-2.0 Apache License 2.0
  * @link      https://github.com/serlo-org/api.serlo.org for the canonical source repository
  */
@@ -83,7 +83,12 @@ export const resolvers: InterfaceResolvers<'ThreadAware'> &
       return resolveConnection({
         nodes: threads,
         payload: { ...input, first, after },
-        createCursor: (node) => node.commentPayloads[0].date,
+        createCursor: (node) => {
+          const comments = node.commentPayloads
+          const latestComment = comments[comments.length - 1]
+
+          return latestComment.date
+        },
       })
     },
   },