+++ scripts/changelog.ts
@@ -1,7 +1,7 @@
 /**
  * This file is part of Serlo.org API
  *
- * Copyright (c) 2020-2022 Serlo Education e.V.
+ * Copyright (c) 2020-2023 Serlo Education e.V.
  *
  * Licensed under the Apache License, Version 2.0 (the "License")
  * you may not use this file except in compliance with the License
@@ -15,7 +15,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  *
- * @copyright Copyright (c) 2020-2022 Serlo Education e.V.
+ * @copyright Copyright (c) 2020-2023 Serlo Education e.V.
  * @license   http://www.apache.org/licenses/LICENSE-2.0 Apache License 2.0
  * @link      https://github.com/serlo-org/api.serlo.org for the canonical source repository
  */
@@ -1421,6 +1421,20 @@ async function exec(): Promise<void> {
         date: '2022-11-22',
 
         fixed: ['Fix sorting in `allThreads` query by last reacted #739'],
+      },
+      {
+        tagName: 'v0.47.2',
+        date: '2022-12-07',
+        fixed: ['Fix pagination in `allThreads`'],
+      },
+      {
+        tagName: 'v0.48.0',
+        date: '2023-01-12',
+        added: [
+          'Endpoint for synchronizing users between new and old tables',
+          'Oauth mutations',
+        ],
+        breakingChanges: ['authenticate user using cookie'],
       },
     ],
   })